pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-migrate>=4.1.0",
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "flask-bcrypt>=1.0.1",
    "wtforms>=3.2.1",
    "passlib[bcrypt]>=1.7.4",
    "flask-login>=0.6.3",
    "flask-wtf>=1.2.2",
]
